# Only run expensive setup where needed in Proquest spec