There is some very expensive workflow and sample datea setup
in the Proquest spec that was being run for every test,
even where it wasn't needed.

This PR moves that setup to only the scoped tests that absolutely
require it.  The PR also uses RSpec's :aggregage_failures feature
to group mulitple small examples into the same test so that fewer
total factory objects need to be created (also somewhat expensive).